replacement: Move neutral payload helpers into core
Replacement handling currently lives in the batch package, where the source
view builder also owns the byte payload model used by other layers.

The project lacks a neutral home for replacement payloads that are shared
by CLI parsing, live staging, candidate preview, and batch source
rewriting. That makes unrelated layers depend on batch internals for common
byte normalization.

This commit begins separating those responsibilities by moving the neutral
payload types and byte-line helpers into core while leaving the batch
source view builder in the batch package.

The next commit will move the replacement tests to the neutral module
before production callers adopt it.
